transpiler: support IfStmt representation of Clang 9.0
Prior to this commit, the transpileIfStmt logic of c2go relied on the
assumption that if-statements always have 4 (or 5) children. With recent
versions of Clang, this assumption no longer holds.

Therefore, this commit relaxes the assumption by simply trimming the
first two child nodes if the IfStmt has 5 children, and the first child
node if the IfStmt has 4 children, respectively.

The later logic is adjusted to always use children[0] for the
conditional node, children[1] for the if-body node and optionally
children[2] for the else-body node (if present).

Fixes #833.